Move security config to type_cds
